@@ -6,8 +6,8 @@ ext {
 
     android = [
             minSdkVersion    : 14,
-            targetSdkVersion : 30,
-            compileSdkVersion: 30,
+            targetSdkVersion : 31,
+            compileSdkVersion: 31,
             buildToolsVersion: "30.0.3",
     ]
 
