@@ -1,6 +1,7 @@
 package com.clj.fastble.bluetooth;
 
 
+import android.annotation.SuppressLint;
 import android.bluetooth.BluetoothDevice;
 import android.os.Build;
 
@@ -25,112 +26,132 @@ public class MultipleBluetoothController {
         bleTempHashMap = new HashMap<>();
     }
 
-    public synchronized BleBluetooth buildConnectingBle(BleDevice bleDevice) {
-        BleBluetooth bleBluetooth = new BleBluetooth(bleDevice);
-        if (!bleTempHashMap.containsKey(bleBluetooth.getDeviceKey())) {
-            bleTempHashMap.put(bleBluetooth.getDeviceKey(), bleBluetooth);
+    public BleBluetooth buildConnectingBle(BleDevice bleDevice) {
+        synchronized (this) {
+            if (!bleTempHashMap.containsKey(bleDevice.getKey())) {
+                BleBluetooth bleBluetooth = new BleBluetooth(bleDevice);
+                bleTempHashMap.put(bleDevice.getKey(), bleBluetooth);
+                return bleBluetooth;
+            }
+            return bleTempHashMap.get(bleDevice.getKey());
         }
-        return bleBluetooth;
     }
 
-    public synchronized void removeConnectingBle(BleBluetooth bleBluetooth) {
-        if (bleBluetooth == null) {
-            return;
-        }
-        if (bleTempHashMap.containsKey(bleBluetooth.getDeviceKey())) {
+    public void removeConnectingBle(BleBluetooth bleBluetooth) {
+        synchronized (this) {
+            if (bleBluetooth == null) {
+                return;
+            }
             bleTempHashMap.remove(bleBluetooth.getDeviceKey());
         }
     }
 
-    public synchronized void addBleBluetooth(BleBluetooth bleBluetooth) {
-        if (bleBluetooth == null) {
-            return;
-        }
-        if (!bleLruHashMap.containsKey(bleBluetooth.getDeviceKey())) {
+    public void addBleBluetooth(BleBluetooth bleBluetooth) {
+        synchronized (this) {
+            if (bleBluetooth == null) {
+                return;
+            }
+            if (bleLruHashMap.containsKey(bleBluetooth.getDeviceKey())) {
+                return;
+            }
             bleLruHashMap.put(bleBluetooth.getDeviceKey(), bleBluetooth);
         }
     }
 
-    public synchronized void removeBleBluetooth(BleBluetooth bleBluetooth) {
-        if (bleBluetooth == null) {
-            return;
-        }
-        if (bleLruHashMap.containsKey(bleBluetooth.getDeviceKey())) {
+    public void removeBleBluetooth(BleBluetooth bleBluetooth) {
+        synchronized (this) {
+            if (bleBluetooth == null) {
+                return;
+            }
             bleLruHashMap.remove(bleBluetooth.getDeviceKey());
         }
     }
 
-    public synchronized boolean isContainDevice(BleDevice bleDevice) {
+    public boolean isContainDevice(BleDevice bleDevice) {
         return bleDevice != null && bleLruHashMap.containsKey(bleDevice.getKey());
     }
 
-    public synchronized boolean isContainDevice(BluetoothDevice bluetoothDevice) {
+    @SuppressLint("MissingPermission")
+    public boolean isContainDevice(BluetoothDevice bluetoothDevice) {
         return bluetoothDevice != null && bleLruHashMap.containsKey(bluetoothDevice.getName() + bluetoothDevice.getAddress());
     }
 
-    public synchronized BleBluetooth getBleBluetooth(BleDevice bleDevice) {
-        if (bleDevice != null) {
-            if (bleLruHashMap.containsKey(bleDevice.getKey())) {
+    public BleBluetooth getBleBluetooth(BleDevice bleDevice) {
+        synchronized (this) {
+            if (bleDevice != null) {
                 return bleLruHashMap.get(bleDevice.getKey());
             }
-        }
-        return null;
-    }
-
-    public synchronized void disconnect(BleDevice bleDevice) {
-        if (isContainDevice(bleDevice)) {
-            getBleBluetooth(bleDevice).disconnect();
+            return null;
         }
     }
 
-    public synchronized void disconnectAllDevice() {
-        for (Map.Entry<String, BleBluetooth> stringBleBluetoothEntry : bleLruHashMap.entrySet()) {
-            stringBleBluetoothEntry.getValue().disconnect();
-        }
-        bleLruHashMap.clear();
+    public void disconnect(BleDevice bleDevice) {
+        getBleBluetooth(bleDevice).disconnect();
     }
 
-    public synchronized void destroy() {
-        for (Map.Entry<String, BleBluetooth> stringBleBluetoothEntry : bleLruHashMap.entrySet()) {
-            stringBleBluetoothEntry.getValue().destroy();
-        }
-        bleLruHashMap.clear();
-        for (Map.Entry<String, BleBluetooth> stringBleBluetoothEntry : bleTempHashMap.entrySet()) {
-            stringBleBluetoothEntry.getValue().destroy();
-        }
-        bleTempHashMap.clear();
-    }
-
-    public synchronized List<BleBluetooth> getBleBluetoothList() {
-        List<BleBluetooth> bleBluetoothList = new ArrayList<>(bleLruHashMap.values());
-        Collections.sort(bleBluetoothList, new Comparator<BleBluetooth>() {
-            @Override
-            public int compare(BleBluetooth lhs, BleBluetooth rhs) {
-                return lhs.getDeviceKey().compareToIgnoreCase(rhs.getDeviceKey());
+    public void disconnectAllDevice() {
+        synchronized (this) {
+            for (Map.Entry<String, BleBluetooth> stringBleBluetoothEntry : bleLruHashMap.entrySet()) {
+                stringBleBluetoothEntry.getValue().disconnect();
             }
-        });
-        return bleBluetoothList;
+            bleLruHashMap.clear();
+        }
     }
 
-    public synchronized List<BleDevice> getDeviceList() {
-        refreshConnectedDevice();
-        List<BleDevice> deviceList = new ArrayList<>();
-        for (BleBluetooth BleBluetooth : getBleBluetoothList()) {
-            if (BleBluetooth != null) {
-                deviceList.add(BleBluetooth.getDevice());
+    public void destroy(String mac) {
+        synchronized (this) {
+            for (Map.Entry<String, BleBluetooth> stringBleBluetoothEntry : bleLruHashMap.entrySet()) {
+                if (mac != null) {
+                    if (stringBleBluetoothEntry.getValue().getDevice().getMac().equals(mac)) {
+                        stringBleBluetoothEntry.getValue().destroy();
+                        bleLruHashMap.remove(stringBleBluetoothEntry.getKey());
+                    }
+                } else {
+                    stringBleBluetoothEntry.getValue().destroy();
+                    bleLruHashMap.remove(stringBleBluetoothEntry.getKey());
+                }
+            }
+            for (Map.Entry<String, BleBluetooth> stringBleBluetoothEntry : bleTempHashMap.entrySet()) {
+                if (mac != null) {
+                    if (stringBleBluetoothEntry.getValue().getDevice().getMac().equals(mac)) {
+                        stringBleBluetoothEntry.getValue().destroy();
+                        bleTempHashMap.remove(stringBleBluetoothEntry.getKey());
+                    }
+                } else {
+                    stringBleBluetoothEntry.getValue().destroy();
+                    bleTempHashMap.remove(stringBleBluetoothEntry.getKey());
+                }
             }
         }
-        return deviceList;
+    }
+
+    public List<BleBluetooth> getBleBluetoothList() {
+        synchronized (this) {
+            List<BleBluetooth> bleBluetoothList = new ArrayList<>(bleLruHashMap.values());
+            Collections.sort(bleBluetoothList, (lhs, rhs) -> lhs.getDeviceKey().compareToIgnoreCase(rhs.getDeviceKey()));
+            return bleBluetoothList;
+        }
+    }
+
+    public List<BleDevice> getDeviceList() {
+        synchronized (this) {
+            refreshConnectedDevice();
+            List<BleDevice> deviceList = new ArrayList<>();
+            for (BleBluetooth BleBluetooth : getBleBluetoothList()) {
+                if (BleBluetooth != null) {
+                    deviceList.add(BleBluetooth.getDevice());
+                }
+            }
+            return deviceList;
+        }
     }
 
     public void refreshConnectedDevice() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
-            List<BleBluetooth> bluetoothList = getBleBluetoothList();
-            for (int i = 0; bluetoothList != null && i < bluetoothList.size(); i++) {
-                BleBluetooth bleBluetooth = bluetoothList.get(i);
-                if (!BleManager.getInstance().isConnected(bleBluetooth.getDevice())) {
-                    removeBleBluetooth(bleBluetooth);
-                }
+        List<BleBluetooth> bluetoothList = getBleBluetoothList();
+        for (int i = 0; bluetoothList != null && i < bluetoothList.size(); i++) {
+            BleBluetooth bleBluetooth = bluetoothList.get(i);
+            if (!BleManager.getInstance().isConnected(bleBluetooth.getDevice())) {
+                removeBleBluetooth(bleBluetooth);
             }
         }
     }
