@@ -2,7 +2,7 @@ package com.clj.fastble.callback;
 
 import com.clj.fastble.data.BleDevice;
 
-public interface BleScanPresenterImp {
+public interface BleScanListener {
 
     void onScanStarted(boolean success);
 
