@@ -10,7 +10,7 @@ public class BleMsg {
     // Connect
     public static final int MSG_CONNECT_FAIL = 0x01;
     public static final int MSG_DISCONNECTED = 0x02;
-    public static final int MSG_RECONNECT = 0x03;
+//    public static final int MSG_RECONNECT = 0x03;
     public static final int MSG_DISCOVER_SERVICES = 0x04;
     public static final int MSG_DISCOVER_FAIL = 0x05;
     public static final int MSG_DISCOVER_SUCCESS = 0x06;
