@@ -1,15 +1,14 @@
 package com.clj.fastble.scan;
 
 
+import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
 import android.os.Build;
-import android.os.Handler;
-import android.os.Looper;
 
 import com.clj.fastble.BleManager;
 import com.clj.fastble.callback.BleScanAndConnectCallback;
 import com.clj.fastble.callback.BleScanCallback;
-import com.clj.fastble.callback.BleScanPresenterImp;
+import com.clj.fastble.callback.BleScanListener;
 import com.clj.fastble.data.BleDevice;
 import com.clj.fastble.data.BleScanState;
 import com.clj.fastble.utils.BleLog;
@@ -34,7 +33,7 @@ public class BleScanner {
 
         @Override
         public void onScanStarted(boolean success) {
-            BleScanPresenterImp callback = mBleScanPresenter.getBleScanPresenterImp();
+            BleScanListener callback = getBleScanListener();
             if (callback != null) {
                 callback.onScanStarted(success);
             }
@@ -42,23 +41,21 @@ public class BleScanner {
 
         @Override
         public void onLeScan(BleDevice bleDevice) {
+            BleScanListener callback = mBleScanPresenter.getBleScanListener();
             if (mBleScanPresenter.ismNeedConnect()) {
-                BleScanAndConnectCallback callback = (BleScanAndConnectCallback)
-                        mBleScanPresenter.getBleScanPresenterImp();
                 if (callback != null) {
-                    callback.onLeScan(bleDevice);
+                    ((BleScanAndConnectCallback) callback).onLeScan(bleDevice);
                 }
             } else {
-                BleScanCallback callback = (BleScanCallback) mBleScanPresenter.getBleScanPresenterImp();
                 if (callback != null) {
-                    callback.onLeScan(bleDevice);
+                    ((BleScanCallback) callback).onLeScan(bleDevice);
                 }
             }
         }
 
         @Override
         public void onScanning(BleDevice result) {
-            BleScanPresenterImp callback = mBleScanPresenter.getBleScanPresenterImp();
+            BleScanListener callback = mBleScanPresenter.getBleScanListener();
             if (callback != null) {
                 callback.onScanning(result);
             }
@@ -66,31 +63,25 @@ public class BleScanner {
 
         @Override
         public void onScanFinished(List<BleDevice> bleDeviceList) {
+            final BleScanListener callback =
+                    mBleScanPresenter.getBleScanListener();
             if (mBleScanPresenter.ismNeedConnect()) {
-                final BleScanAndConnectCallback callback = (BleScanAndConnectCallback)
-                        mBleScanPresenter.getBleScanPresenterImp();
                 if (bleDeviceList == null || bleDeviceList.size() < 1) {
                     if (callback != null) {
-                        callback.onScanFinished(null);
+                        ((BleScanAndConnectCallback) callback).onScanFinished(null);
                     }
                 } else {
                     if (callback != null) {
-                        callback.onScanFinished(bleDeviceList.get(0));
+                        ((BleScanAndConnectCallback) callback).onScanFinished(bleDeviceList.get(0));
                     }
-                    final List<BleDevice> list = bleDeviceList;
-                    new Handler(Looper.getMainLooper()).postDelayed(new Runnable() {
-                        @Override
-                        public void run() {
-                            BleManager.getInstance().connect(list.get(0), callback);
-                        }
-                    }, 100);
+                    BleManager.getInstance().connect(bleDeviceList.get(0), ((BleScanAndConnectCallback) callback));
                 }
             } else {
-                BleScanCallback callback = (BleScanCallback) mBleScanPresenter.getBleScanPresenterImp();
                 if (callback != null) {
-                    callback.onScanFinished(bleDeviceList);
+                    ((BleScanCallback) callback).onScanFinished(bleDeviceList);
                 }
             }
+            mBleScanPresenter.setCancelBleScanListener();
         }
     };
 
@@ -106,29 +97,39 @@ public class BleScanner {
         startLeScan(serviceUuids, names, mac, fuzzy, true, timeOut, callback);
     }
 
-    private synchronized void startLeScan(UUID[] serviceUuids, String[] names, String mac, boolean fuzzy,
-                                          boolean needConnect, long timeOut, BleScanPresenterImp imp) {
-
-        if (mBleScanState != BleScanState.STATE_IDLE) {
-            BleLog.w("scan action already exists, complete the previous scan action first");
-            if (imp != null) {
-                imp.onScanStarted(false);
+    @SuppressLint("MissingPermission")
+    private void startLeScan(UUID[] serviceUuids, String[] names, String mac, boolean fuzzy,
+                             boolean needConnect, long timeOut, BleScanListener imp) {
+        synchronized (this) {
+            if (mBleScanState != BleScanState.STATE_IDLE) {
+                BleLog.w("scan action already exists, complete the previous scan action first");
+                if (imp != null) {
+                    imp.onScanStarted(false);
+                }
+                return;
             }
-            return;
+
+            mBleScanPresenter.prepare(names, mac, fuzzy, needConnect, timeOut, imp);
+
+            boolean success = BleManager.getInstance().getBluetoothAdapter()
+                    .startLeScan(serviceUuids, mBleScanPresenter);
+            mBleScanState = success ? BleScanState.STATE_SCANNING : BleScanState.STATE_IDLE;
+            mBleScanPresenter.notifyScanStarted(success);
         }
-
-        mBleScanPresenter.prepare(names, mac, fuzzy, needConnect, timeOut, imp);
-
-        boolean success = BleManager.getInstance().getBluetoothAdapter()
-                .startLeScan(serviceUuids, mBleScanPresenter);
-        mBleScanState = success ? BleScanState.STATE_SCANNING : BleScanState.STATE_IDLE;
-        mBleScanPresenter.notifyScanStarted(success);
     }
 
-    public synchronized void stopLeScan() {
-        BleManager.getInstance().getBluetoothAdapter().stopLeScan(mBleScanPresenter);
-        mBleScanState = BleScanState.STATE_IDLE;
-        mBleScanPresenter.notifyScanStopped();
+    @SuppressLint("MissingPermission")
+    public void stopLeScan(boolean isCallbackScanFinish) {
+        synchronized (this) {
+            BleManager.getInstance().getBluetoothAdapter().stopLeScan(mBleScanPresenter);
+            mBleScanState = BleScanState.STATE_IDLE;
+            if (!isCallbackScanFinish) {
+                mBleScanPresenter.removeHandlerMsg();
+                mBleScanPresenter.setCancelBleScanListener();
+                return;
+            }
+            mBleScanPresenter.notifyScanStopped();
+        }
     }
 
     public BleScanState getScanState() {
